+++ graphene_mongo/asynchronous/fields.py
@@ -1,71 +1,53 @@
 from __future__ import absolute_import
 
-import logging
-from collections import OrderedDict
-from functools import partial, reduce
+from functools import partial
 from itertools import filterfalse
+from typing import Coroutine
 
 import bson
-import graphene
-import mongoengine
-import pymongo
 from bson import DBRef, ObjectId
+import graphene
 from graphene import Context
 from graphene.relay import ConnectionField
-from graphene.types.argument import to_arguments
-from graphene.types.dynamic import Dynamic
-from graphene.types.structures import Structure
-from graphene.types.utils import get_type
 from graphene.utils.str_converters import to_snake_case
 from graphql import GraphQLResolveInfo
 from graphql_relay import cursor_to_offset, from_global_id
-from mongoengine import QuerySet
+import mongoengine
+from mongoengine import AsyncQuerySet, QuerySet
 from promise import Promise
+import pymongo
 from pymongo.errors import OperationFailure
 
-from .advanced_types import (
-    FileFieldType,
-    MultiPolygonFieldType,
-    PointFieldInputType,
-    PointFieldType,
-    PolygonFieldType,
-)
-from .converter import MongoEngineConversionError, convert_mongoengine_field
-from .registry import get_global_registry
-from .utils import (
+from ..synchronous.fields import MongoengineConnectionField
+from ..base.registry import get_global_async_registry
+from ..base.utils import (
     ExecutorEnum,
     connection_from_iterables,
     find_skip_and_limit,
+    get_document,
     get_model_reference_fields,
     get_query_fields,
     has_page_info,
-    get_document,
 )
 
 PYMONGO_VERSION = tuple(pymongo.version_tuple[:2])
 
 
-class MongoengineConnectionField(ConnectionField):
+class AsyncMongoengineConnectionField(MongoengineConnectionField):
     def __init__(self, type, *args, **kwargs):
-        get_queryset = kwargs.pop("get_queryset", None)
-        if get_queryset:
-            assert callable(get_queryset), (
-                "Attribute `get_queryset` on {} must be callable.".format(self)
-            )
-        self._get_queryset = get_queryset
-        super(MongoengineConnectionField, self).__init__(type, *args, **kwargs)
+        super(AsyncMongoengineConnectionField, self).__init__(type, *args, **kwargs)
 
     @property
-    def executor(self) -> ExecutorEnum:
-        return ExecutorEnum.SYNC
+    def executor(self):
+        return ExecutorEnum.ASYNC
 
     @property
     def type(self):
-        from .types import MongoengineObjectType
+        from .types import AsyncMongoengineObjectType
 
         _type = super(ConnectionField, self).type
-        assert issubclass(_type, MongoengineObjectType), (
-            "MongoengineConnectionField only accepts MongoengineObjectType types"
+        assert issubclass(_type, AsyncMongoengineObjectType), (
+            "AsyncMongoengineConnectionField only accepts AsyncMongoengineObjectType types"
         )
         assert _type._meta.connection, "The type {} doesn't have a connection".format(
             _type.__name__
@@ -73,211 +55,12 @@ class MongoengineConnectionField(ConnectionField):
         return _type._meta.connection
 
     @property
-    def node_type(self):
-        return self.type._meta.node
-
-    @property
-    def model(self):
-        return self.node_type._meta.model
-
-    @property
-    def order_by(self):
-        return self.node_type._meta.order_by
-
-    @property
-    def required_fields(self):
-        return tuple(set(self.node_type._meta.required_fields + self.node_type._meta.only_fields))
-
-    @property
-    def registry(self):
-        return getattr(self.node_type._meta, "registry", get_global_registry())
-
-    @property
-    def args(self):
-        _field_args = self.field_args
-        _advance_args = self.advance_args
-        _filter_args = self.filter_args
-        _extended_args = self.extended_args
-        if self._type._meta.non_filter_fields:
-            for _field in self._type._meta.non_filter_fields:
-                if _field in _field_args:
-                    _field_args.pop(_field)
-                if _field in _advance_args:
-                    _advance_args.pop(_field)
-                if _field in _filter_args:
-                    _filter_args.pop(_field)
-                if _field in _extended_args:
-                    _filter_args.pop(_field)
-        extra_args = dict(
-            dict(dict(_field_args, **_advance_args), **_filter_args), **_extended_args
-        )
-
-        for key in list(self._base_args.keys()):
-            extra_args.pop(key, None)
-        return to_arguments(self._base_args or OrderedDict(), extra_args)
-
-    @args.setter
-    def args(self, args):
-        self._base_args = args
-
-    def _field_args(self, items):
-        def is_filterable(k):
-            """
-            Remove complex columns from input args at this moment.
-
-            Args:
-                k (str): field name.
-            Returns:
-                bool
-            """
-            if hasattr(self.fields[k].type, "_sdl"):
-                return False
-            if not hasattr(self.model, k):
-                return False
-            else:
-                # else section is a patch for federated field error
-                field_ = self.fields[k]
-                type_ = field_.type
-                while hasattr(type_, "of_type"):
-                    type_ = type_.of_type
-                if hasattr(type_, "_sdl") and "@key" in type_._sdl:
-                    return False
-            if isinstance(getattr(self.model, k), property):
-                return False
-            try:
-                converted = convert_mongoengine_field(
-                    getattr(self.model, k), self.registry, self.executor
-                )
-            except MongoEngineConversionError:
-                return False
-            if isinstance(converted, (ConnectionField, Dynamic)):
-                return False
-            if callable(getattr(converted, "type", None)) and isinstance(
-                converted.type(),
-                (
-                    FileFieldType,
-                    PointFieldType,
-                    MultiPolygonFieldType,
-                    graphene.Union,
-                    PolygonFieldType,
-                ),
-            ):
-                return False
-
-            if isinstance(converted, (graphene.List)):
-                sub_type = getattr(converted, "_of_type", None)
-                if hasattr(sub_type, "of_type"):  # graphene.NonNull
-                    sub_type = sub_type.of_type
-                if issubclass(sub_type, graphene.Union) or issubclass(
-                    sub_type, graphene.ObjectType
-                ):
-                    return False
-
-            # below if condition: workaround for DB filterable field redefined as custom graphene type
-            if (
-                hasattr(field_, "type")
-                and hasattr(converted, "type")
-                and converted.type != field_.type
-            ):
-                return False
-            return True
-
-        def get_filter_type(_type):
-            """
-            Returns the scalar type.
-            """
-            if isinstance(_type, Structure):
-                return get_filter_type(_type.of_type)
-            return _type()
-
-        return {k: get_filter_type(v.type) for k, v in items if is_filterable(k)}
-
-    @property
-    def field_args(self):
-        return self._field_args(self.fields.items())
-
-    @property
-    def filter_args(self):
-        filter_args = dict()
-        if self._type._meta.filter_fields:
-            for field, filter_collection in self._type._meta.filter_fields.items():
-                for each in filter_collection:
-                    if str(self._type._meta.fields[field].type) in (
-                        "PointFieldType",
-                        "PointFieldType!",
-                    ):
-                        if each == "max_distance":
-                            filter_type = graphene.Int
-                        else:
-                            filter_type = PointFieldInputType
-                    else:
-                        filter_type = getattr(
-                            graphene,
-                            str(self._type._meta.fields[field].type).replace("!", ""),
-                        )
-                    # handle special cases
-                    advanced_filter_types = {
-                        "in": graphene.List(filter_type),
-                        "nin": graphene.List(filter_type),
-                        "all": graphene.List(filter_type),
-                    }
-                    filter_type = advanced_filter_types.get(each, filter_type)
-                    filter_args[field + "__" + each] = graphene.Argument(type_=filter_type)
-        return filter_args
-
-    @property
-    def advance_args(self):
-        def get_advance_field(r, kv):
-            field = kv[1]
-            mongo_field = getattr(self.model, kv[0], None)
-            if isinstance(mongo_field, mongoengine.PointField):
-                r.update({kv[0]: graphene.Argument(PointFieldInputType)})
-                return r
-            if isinstance(
-                mongo_field,
-                (
-                    mongoengine.ReferenceField,
-                    mongoengine.GenericReferenceField,
-                ),
-            ):
-                r.update({kv[0]: graphene.ID()})
-                return r
-            if isinstance(mongo_field, mongoengine.GenericReferenceField):
-                r.update({kv[0]: graphene.ID()})
-                return r
-            if callable(getattr(field, "get_type", None)):
-                _type = field.get_type()
-                if _type:
-                    node = (
-                        _type.type._meta
-                        if hasattr(_type.type, "_meta")
-                        else _type.type._of_type._meta
-                    )
-                    if "id" in node.fields and not issubclass(
-                        node.model, (mongoengine.EmbeddedDocument,)
-                    ):
-                        r.update({kv[0]: node.fields["id"]._type.of_type()})
-
-            return r
-
-        return reduce(get_advance_field, self.fields.items(), {})
-
-    @property
-    def extended_args(self):
-        args = OrderedDict()
-        for k, each in self.fields.items():
-            if hasattr(each.type, "_sdl"):
-                args.update({k: graphene.ID()})
-        return args
-
-    @property
     def fields(self):
-        self._type = get_type(self._type)
-        return self._type._meta.fields
+        return super(AsyncMongoengineConnectionField, self).fields
 
     def get_queryset(
         self, model, info, required_fields=None, skip=None, limit=None, **args
-    ) -> QuerySet:
+    ) -> AsyncQuerySet:
         if required_fields is None:
             required_fields = list()
 
@@ -320,23 +103,31 @@ class MongoengineConnectionField(ConnectionField):
 
         if self._get_queryset:
             queryset_or_filters = self._get_queryset(model, info, **args)
-            if isinstance(queryset_or_filters, mongoengine.QuerySet):
+            if isinstance(queryset_or_filters, mongoengine.AsyncQuerySet):
                 return queryset_or_filters
+            elif isinstance(queryset_or_filters, mongoengine.QuerySet):
+                raise TypeError(
+                    "AsyncMongoengineConnectionField only accepts AsyncQuerySet in get_queryset(...)"
+                )
             else:
                 args.update(queryset_or_filters)
         if limit is not None:
             return (
-                model.objects(**args)
+                model.aobjects(**args)
                 .only(*required_fields)
                 .order_by(self.order_by)
                 .skip(skip if skip else 0)
                 .limit(limit)
             )
         elif skip is not None:
-            return model.objects(**args).only(*required_fields).order_by(self.order_by).skip(skip)
-        return model.objects(**args).only(*required_fields).order_by(self.order_by)
+            return model.aobjects(**args).only(*required_fields).order_by(self.order_by).skip(skip)
+        return model.aobjects(**args).only(*required_fields).order_by(self.order_by)
 
-    def default_resolver(self, _root, info, required_fields=None, resolved=None, **args):
+    @property
+    def registry(self):
+        return getattr(self.node_type._meta, "registry", get_global_async_registry())
+
+    async def default_resolver(self, _root, info, required_fields=None, resolved=None, **args):
         if required_fields is None:
             required_fields = list()
         args = args or {}
@@ -381,14 +172,14 @@ class MongoengineConnectionField(ConnectionField):
         if resolved is not None:
             items = resolved
 
-            if isinstance(items, QuerySet):
+            if isinstance(items, AsyncQuerySet):
                 try:
                     if last is not None:
-                        count = items.count(with_limit_and_skip=False)
+                        count = await items.count(with_limit_and_skip=False)
                     else:
                         count = None
                 except OperationFailure:
-                    count = len(items)
+                    count = len(await items.to_list())
             else:
                 count = len(items)
 
@@ -396,11 +187,18 @@ class MongoengineConnectionField(ConnectionField):
                 first=first, last=last, after=after, before=before, count=count
             )
 
-            if isinstance(items, QuerySet):
+            if isinstance(items, AsyncQuerySet):
                 if limit:
-                    _base_query: QuerySet = items.skip(skip)
-                    items = _base_query.limit(limit)
-                    has_next_page = len(_base_query.skip(skip + limit).only("id").limit(1)) != 0
+                    _base_query: AsyncQuerySet = await items.skip(skip)
+                    items = await _base_query.limit(limit)
+                    has_next_page = (
+                        (
+                            len(await _base_query.skip(skip + limit).only("id").limit(1).to_list())
+                            != 0
+                        )
+                        if requires_page_info
+                        else False
+                    )
                 elif skip:
                     items = items.skip(skip)
             else:
@@ -412,7 +210,7 @@ class MongoengineConnectionField(ConnectionField):
                     )
                 elif skip:
                     items = items[skip:]
-            iterables = list(items)
+            iterables = await items.to_list()
             list_length = len(iterables)
 
         elif callable(getattr(self.model, "objects", None)):
@@ -427,6 +225,7 @@ class MongoengineConnectionField(ConnectionField):
                     elif skip:
                         args["pk__in"] = args["pk__in"][skip:]
                     iterables = self.get_queryset(self.model, info, required_fields, **args)
+                    iterables = await iterables.to_list()
                 else:
                     # If there is no ids to fetch, No need of DB call
                     iterables = []
@@ -441,7 +240,7 @@ class MongoengineConnectionField(ConnectionField):
             elif (
                 _root is None
                 or args
-                or isinstance(getattr(_root, field_name, []), MongoengineConnectionField)
+                or isinstance(getattr(_root, field_name, []), AsyncMongoengineConnectionField)
             ):
                 args_copy = args.copy()
                 for key in args.copy():
@@ -463,14 +262,7 @@ class MongoengineConnectionField(ConnectionField):
                         if getattr(args_copy[key], "value", None):
                             args_copy[key] = args_copy[key].value
 
-                if PYMONGO_VERSION >= (3, 7):
-                    if hasattr(self.model, "_meta") and "db_alias" in self.model._meta:
-                        db = mongoengine.get_db(self.model._meta["db_alias"])
-                    else:
-                        db = mongoengine.get_db()
-                    count = db[self.model._get_collection_name()].count_documents(args_copy)
-                else:
-                    count = self.model.objects(args_copy).count()
+                count = await self.model.aobjects(**args_copy).count()
                 if count != 0:
                     skip, limit = find_skip_and_limit(
                         first=first, after=after, last=last, before=before, count=count
@@ -478,6 +270,7 @@ class MongoengineConnectionField(ConnectionField):
                     iterables = self.get_queryset(
                         self.model, info, required_fields, skip, limit, **args
                     )
+                    iterables = await iterables.to_list()
                     list_length = len(iterables)
                     if isinstance(info, GraphQLResolveInfo):
                         if not info.context:
@@ -500,15 +293,16 @@ class MongoengineConnectionField(ConnectionField):
             elif skip:
                 items = items[skip:]
             iterables = items
+            iterables = await iterables.to_list()
             list_length = len(iterables)
 
-        if count:
+        if requires_page_info and count:
             has_next_page = (
                 True
                 if (0 if limit is None else limit) + (0 if skip is None else skip) < count
                 else False
             )
-        has_previous_page = True if skip else False
+        has_previous_page = True if requires_page_info and skip else False
 
         connection = connection_from_iterables(
             edges=iterables,
@@ -523,7 +317,7 @@ class MongoengineConnectionField(ConnectionField):
         connection.list_length = list_length
         return connection
 
-    def chained_resolver(self, resolver, is_partial, root, info, **args):
+    async def chained_resolver(self, resolver, is_partial, root, info, **args):
         for key, value in dict(args).items():
             if value is None:
                 del args[key]
@@ -547,7 +341,7 @@ class MongoengineConnectionField(ConnectionField):
                 connection_fields = [
                     field
                     for field in self.fields
-                    if isinstance(self.fields[field], MongoengineConnectionField)
+                    if isinstance(self.fields[field], AsyncMongoengineConnectionField)
                 ]
 
                 def filter_connection(x):
@@ -573,15 +367,18 @@ class MongoengineConnectionField(ConnectionField):
 
             # XXX: Filter nested args
             resolved = resolver(root, info, **args)
-
+            if isinstance(resolved, Coroutine):
+                resolved = await resolved
             if resolved is not None:
+                # if isinstance(resolved, Coroutine):
+                #     resolved = await resolved
                 if isinstance(resolved, list):
                     if resolved == list():
                         return resolved
                     elif not isinstance(resolved[0], DBRef):
                         return resolved
                     else:
-                        return self.default_resolver(root, info, required_fields, **args_copy)
+                        return await self.default_resolver(root, info, required_fields, **args_copy)
                 elif isinstance(resolved, QuerySet):
                     args.update(resolved._query)
                     args_copy = args.copy()
@@ -612,40 +409,32 @@ class MongoengineConnectionField(ConnectionField):
                                 operation = list(arg.keys())[0]
                                 args_copy[arg_name + operation.replace("$", "__")] = arg[operation]
                                 del args_copy[arg_name]
-                    return self.default_resolver(
+
+                    return await self.default_resolver(
                         root, info, required_fields, resolved=resolved, **args_copy
                     )
                 elif isinstance(resolved, Promise):
                     return resolved.value
                 else:
-                    return resolved
+                    return await resolved
 
-        return self.default_resolver(root, info, required_fields, **args)
+        return await self.default_resolver(root, info, required_fields, **args)
 
     @classmethod
-    def connection_resolver(cls, resolver, connection_type, root, info, **args):
+    async def connection_resolver(cls, resolver, connection_type, root, info, **args):
         if root:
             for key, value in root.__dict__.items():
                 if value:
                     try:
                         setattr(root, key, from_global_id(value)[1])
-                    except Exception as error:
-                        logging.debug("Exception Occurred: ", exc_info=error)
-        iterable = resolver(root, info, **args)
+                    except Exception:
+                        pass
+
+        iterable = await resolver(root=root, info=info, **args)
 
         if isinstance(connection_type, graphene.NonNull):
             connection_type = connection_type.of_type
-
         on_resolve = partial(cls.resolve_connection, connection_type, args)
-
         if Promise.is_thenable(iterable):
-            return Promise.resolve(iterable).then(on_resolve)
-
+            iterable = Promise.resolve(iterable).then(on_resolve).value
         return on_resolve(iterable)
-
-    def wrap_resolve(self, parent_resolver):
-        super_resolver = self.resolver or parent_resolver
-        resolver = partial(
-            self.chained_resolver, super_resolver, isinstance(super_resolver, partial)
-        )
-        return partial(self.connection_resolver, resolver, self.type)
