+++ tests/asynchronous/nodes.py
@@ -4,7 +4,7 @@ from graphene.relay import Node
 from .. import models
 from .. import types  # noqa: F401
 from ..models import ProfessorMetadata
-from graphene_mongo.types_async import AsyncMongoengineObjectType
+from graphene_mongo.asynchronous.types import AsyncMongoengineObjectType
 
 
 class PublisherAsyncNode(AsyncMongoengineObjectType):
