+++ tests/synchronous/test_fields.py
@@ -2,7 +2,7 @@ import pytest
 
 from . import nodes
 from graphene_mongo import AsyncMongoengineConnectionField
-from graphene_mongo.fields import MongoengineConnectionField
+from graphene_mongo.synchronous.fields import MongoengineConnectionField
 
 
 def test_article_field_args():
