+++ tests/synchronous/test_relay_query.py
@@ -9,8 +9,8 @@ from graphql_relay.node.node import to_global_id
 
 from .. import models
 from . import nodes
-from graphene_mongo.fields import MongoengineConnectionField
-from graphene_mongo.types import MongoengineObjectType
+from graphene_mongo.synchronous.fields import MongoengineConnectionField
+from graphene_mongo.synchronous.types import MongoengineObjectType
