+++ tests/synchronous/test_utils.py
@@ -2,7 +2,7 @@ import graphene
 
 from .. import types
 from ..models import Article, Child, Reporter
-from graphene_mongo.utils import get_model_fields, get_query_fields, is_valid_mongoengine_model
+from graphene_mongo.base.utils import get_model_fields, get_query_fields, is_valid_mongoengine_model
 
 
 def test_get_model_fields_no_duplication():
